refactor: extract assistant error-parsing helpers from ApiStateFlowExtensions into a presentation-independent utility and update related tests to be deterministic
